refactor account slice
