+++ src/redux/accountSlice.ts
@@ -1,93 +1,11 @@
 import { createSlice, createAsyncThunk } from "@reduxjs/toolkit";
 
-import { getAssetsDetailed } from "../store";
 import { getBurrow } from "../utils";
-import { getBalance, getPortfolio } from "../api";
-import { hasZeroSharesFarmRewards, listToMap, transformAccountFarms } from "./utils";
-import { ChangeMethodsLogic, IBoosterStaking } from "../interfaces";
+import { ChangeMethodsLogic } from "../interfaces";
 import { identifyUser } from "../telemetry";
-
-interface Balance {
-  [tokenId: string]: string;
-}
-
-interface PortfolioAsset {
-  apr: string;
-  balance: string;
-  shares: string;
-}
-
-export interface FarmData {
-  boosted_shares: string;
-  unclaimed_amount: string;
-  asset_farm_reward: {
-    reward_per_day: string;
-    booster_log_base: string;
-    remaining_rewards: string;
-    boosted_shares: string;
-  };
-}
-
-export interface Farm {
-  [reward_token_id: string]: FarmData;
-}
-
-export interface Portfolio {
-  supplied: {
-    [tokenId: string]: PortfolioAsset;
-  };
-  collateral: {
-    [tokenId: string]: PortfolioAsset;
-  };
-  borrowed: {
-    [tokenId: string]: PortfolioAsset;
-  };
-  farms: {
-    supplied: {
-      [tokenId: string]: Farm;
-    };
-    borrowed: {
-      [tokenId: string]: Farm;
-    };
-  };
-  staking: IBoosterStaking;
-  hasNonFarmedAssets: boolean;
-}
-
-type Status = "pending" | "fulfilled" | "rejected" | undefined;
-export interface AccountState {
-  accountId: string;
-  balances: Balance;
-  portfolio: Portfolio;
-  status: Status;
-  fetchedAt: string | undefined;
-  isClaiming: Status;
-}
-
-const initialStaking = {
-  staked_booster_amount: "0",
-  unlock_timestamp: "0",
-  x_booster_amount: "0",
-};
-
-const initialState: AccountState = {
-  accountId: "",
-  balances: {},
-  portfolio: {
-    supplied: {},
-    collateral: {},
-    borrowed: {},
-    farms: {
-      supplied: {},
-      borrowed: {},
-    },
-    staking: initialStaking,
-    hasNonFarmedAssets: false,
-  },
-  status: undefined,
-  isClaiming: undefined,
-  fetchedAt: undefined,
-};
+import { transformAccount } from "../transformers/account";
+import { getAccount } from "../api/get-account";
+import { initialState } from "./accountState";
 
 export const farmClaimAll = createAsyncThunk("account/farmClaimAll", async () => {
   const { call, logicContract } = await getBurrow();
@@ -100,23 +18,15 @@ export const farmClaimAll = createAsyncThunk("account/farmClaimAll", async () =>
 });
 
 export const fetchAccount = createAsyncThunk("account/fetchAccount", async () => {
-  const { account } = await getBurrow();
-  const { accountId } = account;
+  const account = await getAccount().then(transformAccount);
 
-  if (accountId) {
+  if (account?.accountId) {
     const wallet = JSON.parse(
       localStorage.getItem("near-wallet-selector:selectedWalletId") || `"undefined"`,
     );
-    identifyUser(accountId, { wallet });
-    const assets = await getAssetsDetailed();
-    const tokenIds = assets.map((asset) => asset.token_id);
-    const accountBalance = (await account.getAccountBalance()).available;
-    const balances = await Promise.all(tokenIds.map((id) => getBalance(id, accountId)));
-    const portfolio = await getPortfolio(accountId);
-    return { accountId, accountBalance, balances, portfolio, tokenIds };
+    identifyUser(account.accountId, { wallet });
   }
-
-  return undefined;
+  return account;
 });
 
 export const accountSlice = createSlice({
@@ -148,24 +58,14 @@ export const accountSlice = createSlice({
       state.status = action.meta.requestStatus;
       state.fetchedAt = new Date().toString();
       if (!action.payload?.accountId) return;
-      const { accountId, accountBalance, balances, portfolio, tokenIds } = action.payload;
+
+      const { accountId, balances, portfolio } = action.payload;
 
       state.accountId = accountId;
-      state.balances = {
-        ...balances.map((b, i) => ({ [tokenIds[i]]: b })).reduce((a, b) => ({ ...a, ...b }), {}),
-        near: accountBalance,
-      };
+      state.balances = balances;
 
       if (portfolio) {
-        const { booster_staking, supplied, borrowed, collateral, farms } = portfolio;
-        state.portfolio = {
-          supplied: listToMap(supplied),
-          borrowed: listToMap(borrowed),
-          collateral: listToMap(collateral),
-          farms: transformAccountFarms(farms),
-          staking: booster_staking || initialStaking,
-          hasNonFarmedAssets: portfolio["has_non_farmed_assets"] || hasZeroSharesFarmRewards(farms),
-        };
+        state.portfolio = portfolio;
       }
     });
   },
