+++ src/redux/selectors/getAccountRewards.ts
@@ -4,7 +4,7 @@ import { omit } from "lodash";
 import { shrinkToken } from "../../store";
 import { RootState } from "../store";
 import { Asset } from "../assetsSlice";
-import { Farm, FarmData, Portfolio } from "../accountSlice";
+import { Farm, FarmData, Portfolio } from "../accountState";
 import { getStaking } from "./getStaking";
 
 interface IPortfolioReward {
