+++ src/redux/selectors/getNetAPY.ts
@@ -4,7 +4,7 @@ import { RootState } from "../store";
 import { hasAssets, toUsd } from "../utils";
 import { getExtraDailyTotals } from "./getExtraDailyTotals";
 import { AssetsState } from "../assetsSlice";
-import { Portfolio } from "../accountSlice";
+import { Portfolio } from "../accountState";
 
 export const getGains = (
   portfolio: Portfolio,
