+++ src/redux/selectors/getPortfolioAssets.ts
@@ -5,7 +5,7 @@ import { RootState } from "../store";
 import { emptySuppliedAsset, emptyBorrowedAsset, hasAssets, getRewards, toUsd } from "../utils";
 import { shrinkToken, expandToken } from "../../store";
 import { Asset, Assets } from "../assetsSlice";
-import { Farm } from "../accountSlice";
+import { Farm } from "../accountState";
 
 export const getPortfolioRewards = (
   type: "supplied" | "borrowed",
