+++ src/redux/selectors/getWithdrawMaxAmount.ts
@@ -5,7 +5,7 @@ import { RootState } from "../store";
 import { shrinkToken, expandTokenDecimal } from "../../store";
 import { decimalMax, decimalMin } from "../../utils";
 import { Assets } from "../assetsSlice";
-import { Portfolio } from "../accountSlice";
+import { Portfolio } from "../accountState";
 
 const sumReducerDecimal = (sum: Decimal, cur: Decimal) => sum.add(cur);
