+++ src/transformers/asstets.ts
@@ -1,0 +1,3 @@
+export function transformAssets() {
+  return undefined;
+}
